+++ src/annotator/config/test/settings-test.js
@@ -236,40 +236,4 @@ describe('annotator.config.settings', function() {
       });
     });
   });
-
-  describe('#isBrowserExtension', function() {
-    [
-      {
-        url: 'chrome-extension://abcxyz',
-        returns: true,
-      },
-      {
-        url: 'moz-extension://abcxyz',
-        returns: true,
-      },
-      {
-        url: 'ms-browser-extension://abcxyz',
-        returns: true,
-      },
-      {
-        url: 'http://partner.org',
-        returns: false,
-      },
-      {
-        url: 'https://partner.org',
-        returns: false,
-      },
-      // It considers anything not http(s) to be a browser extension.
-      {
-        url: 'ftp://partner.org',
-        returns: true,
-      },
-    ].forEach(function(test) {
-      it('returns ' + test.returns + ' for ' + test.url, function() {
-        assert.equal(
-          settings.isBrowserExtension({app: test.url}),
-          test.returns);
-      });
-    });
-  });
 });
